* WARNATON! - 3 guys nonstop hacking night fixing warnings   - Lot of bugs fixed (mostly unexpected segfaults)   - Do not exit when cannot open grava window (gtk_init_check)   - Add mips/code.c (oops)   - Some code refactoring   - Popularize u8, u16, u32, u64 and addr_t (aka u64)   - Drop runtime debugging messages   - Centralize hasher stuff in hasher/hash.h * Add rasm support to compile from files (-f flag)   - Add support for directives and comments     - ';' char comment separator     - .org, .zero, ... * Import bitcarrier changes to socket.c and socket.h * Make rabin -a identify ELF and PE x86 and ARM architectures * Initial import of socket io plugin to open socket connections   as streams (always growing file)

+++ src/dbg/events.c
@@ -114,7 +114,7 @@ int events_init()
 int events_get()
 {
 	int ret;
-	unsigned long ret2;
+	unsigned int ret2;
 
 #if !__x86_64__
 	ret = ptrace(PTRACE_GETEVENTMSG, ps.pid, NULL, &ret2);
